backend/pyproject.toml
[project]
name = "cord"
version = "0.3.0"
description = "Open source voicechat"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.135,<1",
    "uvicorn[standard]>=0.40,<1",
    "sqlalchemy[asyncio]>=2.0,<3",
    "asyncpg>=0.31,<1",
    "pydantic>=2.12,<3",
    "pydantic-settings>=2.13,<3",
    "cryptography>=46,<47",
    "PyJWT>=2.12,<3",
    "bcrypt>=5.0,<6",
    "passlib>=1.7,<2",
    "python-multipart>=0.0.9",
    "redis[asyncio]>=5,<6",
    "livekit-api>=1,<2",
    "httpx>=0.27,<1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8,<9",
    "pytest-asyncio>=0.24,<1",
    "httpx>=0.27,<1",
    "ruff>=0.8,<1",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.ruff]
target-version = "py314"
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
